refs #4727 fix our clearGroup impact on prev/next

+++ civicrm/custom/php/CRM/Core/BAO/PrevNextCache.php
@@ -436,10 +436,12 @@ WHERE  cacheKey LIKE %1
   // Allow for clearing entries based on groupId for when
   // contacts are added and removed group groups.
   function clearGroup( $groupId ) {
-    CRM_Core_DAO::executeQuery(
-      "DELETE FROM civicrm_prevnext_cache WHERE cacheKey LIKE %1",
-      array(1=>array("%_$groupId", 'String'))
-    );
+    if ( !empty($groupId) ) {
+      CRM_Core_DAO::executeQuery(
+        "DELETE FROM civicrm_prevnext_cache WHERE cacheKey LIKE %1",
+        array(1=>array("%_$groupId", 'String'))
+      );
+    }
   }
 }
